% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{acs_cv}
\alias{acs_cv}
\title{Calculate coefficient of variation from an ACS estimate and MOE.}
\usage{
acs_cv(estimate, moe, conf = 0.9)
}
\arguments{
\item{estimate}{Numeric estimate.}

\item{moe}{Numeric MOE.}

\item{conf}{Confidence level associated with \code{moe}.}
}
\value{
Numeric coefficient of variation, using standard error divided by
absolute estimate.
}
\description{
Calculate coefficient of variation from an ACS estimate and MOE.
}
