% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{acs_product}
\alias{acs_product}
\title{Propagate ACS uncertainty for a product.}
\usage{
acs_product(estimate1, moe1, estimate2, moe2, cov = 0, conf = 0.9)
}
\arguments{
\item{estimate1}{First estimate.}

\item{moe1}{MOE for \code{estimate1}.}

\item{estimate2}{Second estimate.}

\item{moe2}{MOE for \code{estimate2}.}

\item{cov}{Covariance between the two estimates on the standard-error scale.}

\item{conf}{Confidence level associated with input and output MOEs.}
}
\value{
A data frame with \code{estimate}, \code{moe}, and \code{se}.
}
\description{
Propagate ACS uncertainty for a product.
}
