% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{acs_ratio}
\alias{acs_ratio}
\title{Propagate ACS uncertainty for a ratio.}
\usage{
acs_ratio(num, num_moe, denom, denom_moe, cov = 0, conf = 0.9)
}
\arguments{
\item{num}{Numerator estimate.}

\item{num_moe}{Numerator MOE.}

\item{denom}{Denominator estimate.}

\item{denom_moe}{Denominator MOE.}

\item{cov}{Numerator-denominator covariance on the standard-error scale.}

\item{conf}{Confidence level associated with input and output MOEs.}
}
\value{
A data frame with \code{estimate}, \code{moe}, and \code{se}.
}
\description{
Propagate ACS uncertainty for a ratio.
}
