% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{moe_ci}
\alias{moe_ci}
\title{Convert an ACS MOE to lower and upper confidence bounds.}
\usage{
moe_ci(estimate, moe, conf_in = 0.9, conf_out = 0.95)
}
\arguments{
\item{estimate}{Numeric estimate.}

\item{moe}{Numeric margin of error.}

\item{conf_in}{Confidence level associated with \code{moe}.}

\item{conf_out}{Desired confidence level for the returned interval.}
}
\value{
A data frame with \code{estimate}, \code{lower}, \code{upper}, and \code{moe}.
}
\description{
Convert an ACS MOE to lower and upper confidence bounds.
}
