% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{se_to_moe}
\alias{se_to_moe}
\title{Convert a standard error to an ACS-style margin of error.}
\usage{
se_to_moe(se, conf = 0.9)
}
\arguments{
\item{se}{Numeric standard error.}

\item{conf}{Desired confidence level.}
}
\value{
Numeric margin of error.
}
\description{
Convert a standard error to an ACS-style margin of error.
}
