% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{moe_to_se}
\alias{moe_to_se}
\title{Convert an ACS margin of error to a standard error.}
\usage{
moe_to_se(moe, conf = 0.9)
}
\arguments{
\item{moe}{Numeric margin of error.}

\item{conf}{Confidence level associated with \code{moe}.}
}
\value{
Numeric standard error.
}
\description{
Convert an ACS margin of error to a standard error.
}
